Split one chapter into two
generation.tex was originally split off from color.tex. It has now itself become too long. Fortunately it is divided into two, the first half dealing with empirical support and pre-Socratic tradition, the second half dealing directly with Aristotle. What was generation.tex is now:

1. light_and_dark.tex
2. generation.tex

Moreover, aristotle.tex has been modified to now include light_and_dark.tex.

Just as I was reluctant to split generation.tex off from color.tex, I was reluctant to further split this material. However, it is already reasonably self-contained and perhaps it is better for the reader to keep the chapters short. The original plan of having chapters on the object, medium, and organ, that is to follow the Aristotelian explanatory taxonomy is now obscured if not lost altogether...